Bug 2033697 - Cleanup MIDI IPC. r=gsvelto
Differential Revision: https://phabricator.services.mozilla.com/D295469

UltraBlame original commit: 4a656380e05edf0f124880f505bc2cfdc1e138c1

+++ dom/midi/MIDIPortParent.cpp
@@ -2,8 +2,6 @@
 
 
 
-
-
 #include "mozilla/dom/MIDIPortParent.h"
 
 #include "mozilla/dom/MIDIPlatformService.h"
@@ -91,9 +89,7 @@ bool MIDIPortParent::SendUpdateStatus(
     }
     mMessageQueue.Clear();
   }
-  return PMIDIPortParent::SendUpdateStatus(
-      static_cast<uint32_t>(mDeviceState),
-      static_cast<uint32_t>(mConnectionState));
+  return PMIDIPortParent::SendUpdateStatus(mDeviceState, mConnectionState);
 }
 
 MIDIPortParent::MIDIPortParent(const MIDIPortInfo& aPortInfo,
